requirements.txt
streamlit
chromadb
sentence-transformers
llama-cpp-python
arxiv
unstructured
numpy